Update template export.
git-svn-id: svn://biometris-nas.wurnet.nl/AMIGA@517 3ae013c2-4042-2242-b34d-985be216b4aa

+++ AmigaPowerAnalysis/AmigaPowerAnalysis.Core/Core/DataAnalysis/AnalysisDataTemplate/AnalysisDataTemplate.cs
@@ -94,7 +94,7 @@ namespace AmigaPowerAnalysis.Core.DataAnalysis {
                     line.Add(factor.ToString());
                 }
                 foreach (var contrast in record.ContrastsPerEndpoint) {
-                    line.Add(string.Format("{0}", contrast));
+                    line.Add(string.Format("{0}", (int)contrast));
                 }
                 stringBuilder.AppendLine(string.Join(separator, line));
             }
